add exib method to hide the logs from server

+++ RC Web/php/db/criar_tb.php
@@ -1,6 +1,17 @@
 <?php
 require('./php/db/conectar_db.php');
 
+$_SERVER['LOG'] = false;
+function exib(string $msg)
+{
+	if($_SERVER['LOG'])
+	{
+		$pi = "<p>";
+		$po = "</p>";
+		echo "$pi $msg $po";
+	}
+}
+
 try {
 	// Tenta estabelecer a conexão com o banco de dados
 	$conectar = new PDO("mysql:host=$host;dbname=$db", $user, $pass);
@@ -8,9 +19,9 @@ try {
 
 	// Verifica se a conexão foi estabelecida com sucesso
 	if ($conectar) {
-		echo "Conexão com o banco de dados estabelecida com sucesso!";
+		exib("Conexão com o banco de dados estabelecida com sucesso!");
 	} else {
-		echo "Falha ao estabelecer a conexão com o banco de dados.";
+		exib("Falha ao estabelecer a conexão com o banco de dados.");
 	}
 } catch (PDOException $e) {
 	// Se houver um erro, exibe a mensagem de erro
@@ -21,17 +32,18 @@ try {
 if ($conectar) {
 
 	try {
-		$pi = "<p>";
-		$po = "</p>";
 		$tb = "CEP";
+		$drop = false;
+		if($drop){
 		$droptb = "DROP TABLE IF EXISTS $db.$tb";
 		$conectar->exec($droptb);
-		echo "$pi DROP TABLE $db.$tb - OK $po";
+		exib("DROP TABLE $db.$tb - OK");
 
 		$tb = "CLIENTES";
 		$droptb = "DROP TABLE IF EXISTS $db.$tb";
 		$conectar->exec($droptb);
-		echo "$pi DROP TABLE $db.$tb - OK $po";
+		exib("DROP TABLE $db.$tb - OK");
+		}
 	} catch (PDOException $erro) {
 		echo "$pi Erro no DROP da tabela $db.$tb" . $erro->getMessage() . $po;
 	}
@@ -54,7 +66,7 @@ if ($conectar) {
 			";
 		//
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. " . $erro->getMessage();
 	}
@@ -68,7 +80,8 @@ if ($conectar) {
 				nome		VARCHAR(70) NOT NULL,
 				email		VARCHAR(30) NOT NULL,
 				cpf_cnpj	VARCHAR(14) NOT NULL,
-				tel			VARCHAR(14) NOT NULL,
+				tel1		VARCHAR(14) NOT NULL,
+				tel2		VARCHAR(14) NULL,
 				cep			CHAR(9)		NOT NULL,
 				end_nro		INT 		NOT NULL,
 				end_cmplto	VARCHAR(45)	NULL,
@@ -76,7 +89,7 @@ if ($conectar) {
 				CONSTRAINT FK_" . $tb . "_CEP FOREIGN KEY (cep) REFERENCES CEP (cod))
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -93,7 +106,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!") ;
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -109,7 +122,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -125,7 +138,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -140,7 +153,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!") ;
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -156,7 +169,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -173,7 +186,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -189,7 +202,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -206,7 +219,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -228,7 +241,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
@@ -261,7 +274,7 @@ if ($conectar) {
 			)
 			ENGINE=InnoDB DEFAULT CHARSET=latin1";
 		$conectar->exec($criatb);
-		echo "<p>CREATE TABLE $tb -> OK!</p>";
+		exib("CREATE TABLE $tb -> OK!");
 	} catch (PDOException $erro) {
 		echo "Tabela $tb. Falhou na criacao. Msg: " . $erro->getMessage();
 	}
